+++ polaris-react/src/components/Box/Box.stories.tsx
@@ -18,71 +18,60 @@ export function Default() {
 export function WithBorders() {
   return (
     <BlockStack gap="400">
-      <Box background="bg" padding="400" borderWidth="1" borderColor="border">
+      <Box
+        background="bg-surface"
+        padding="400"
+        borderWidth="025"
+        borderColor="border"
+      >
         <Text as="p">1px solid border</Text>
       </Box>
       <Box
-        background="bg"
+        background="bg-surface"
         padding="400"
-        borderWidth="1"
+        borderWidth="025"
         borderStyle="dashed"
         borderColor="border-subdued"
       >
         <Text as="p">1px dashed border</Text>
       </Box>
       <Box
-        background="bg"
+        background="bg-surface"
         padding="400"
-        borderWidth="2"
+        borderWidth="050"
         borderColor="border-info"
       >
         <Text as="p">2px solid blue</Text>
       </Box>
       <Box
-        background="bg"
+        background="bg-surface"
         padding="400"
-        borderWidth="3"
-        borderColor="border-success"
-      >
-        <Text as="p">3px solid green</Text>
-      </Box>
-      <Box
-        background="bg"
-        padding="400"
-        borderWidth="4"
+        borderWidth="100"
         borderColor="border-caution"
       >
         <Text as="p">4px solid yellow</Text>
       </Box>
       <Box
-        background="bg"
+        background="bg-surface"
         padding="400"
-        borderWidth="5"
+        borderBlockStartWidth="100"
         borderColor="border-critical"
       >
-        <Text as="p">5px solid red</Text>
+        <Text as="p">border-block-start: 4px solid red</Text>
       </Box>
       <Box
-        background="bg"
+        background="bg-surface"
         padding="400"
-        borderBlockStartWidth="5"
+        borderWidth="100"
+        borderBlockStartWidth="025"
         borderColor="border-critical"
       >
-        <Text as="p">border-block-start: 5px solid red</Text>
-      </Box>
-      <Box
-        background="bg"
-        padding="400"
-        borderWidth="5"
-        borderBlockStartWidth="1"
-        borderColor="border-critical"
-      >
-        <Text as="p">border-width: 5px solid red</Text>
+        <Text as="p">border-width: 4px solid red</Text>
         <Text as="p">border-block-start: 1px solid red</Text>
         <Box
-          background="bg"
+          background="bg-surface"
           padding="400"
-          borderWidth="4"
+          borderWidth="100"
           borderColor="border-caution"
         >
           <Text as="p">border-width: 4px solid yellow</Text>
@@ -93,12 +82,12 @@ export function WithBorders() {
         </Box>
       </Box>
       <Box
-        background="bg"
+        background="bg-surface"
         padding="400"
-        borderWidth="5"
+        borderWidth="100"
         borderColor="transparent"
       >
-        <Text as="p">5px solid transparent</Text>
+        <Text as="p">4px solid transparent</Text>
       </Box>
     </BlockStack>
   );
@@ -107,47 +96,36 @@ export function WithBorders() {
 export function WithOutline() {
   return (
     <BlockStack gap="400">
-      <Box background="bg" padding="400" outlineWidth="1" outlineColor="border">
+      <Box
+        background="bg-surface"
+        padding="400"
+        outlineWidth="025"
+        outlineColor="border"
+      >
         <Icon source={PaintBrushMajor} tone="base" />
       </Box>
       <Box
-        background="bg"
+        background="bg-surface"
         padding="400"
-        outlineWidth="1"
+        outlineWidth="025"
         outlineStyle="dashed"
         outlineColor="border-subdued"
       >
         <Icon source={PaintBrushMajor} tone="base" />
       </Box>
       <Box
-        background="bg"
+        background="bg-surface"
         padding="400"
-        outlineWidth="2"
+        outlineWidth="050"
         outlineColor="border-info"
       >
         <Icon source={PaintBrushMajor} tone="base" />
       </Box>
       <Box
-        background="bg"
+        background="bg-surface"
         padding="400"
-        outlineWidth="3"
-        outlineColor="border-success"
-      >
-        <Icon source={PaintBrushMajor} tone="base" />
-      </Box>
-      <Box
-        background="bg"
-        padding="400"
-        outlineWidth="4"
+        outlineWidth="100"
         outlineColor="border-caution"
-      >
-        <Icon source={PaintBrushMajor} tone="base" />
-      </Box>
-      <Box
-        background="bg"
-        padding="400"
-        outlineWidth="5"
-        outlineColor="border-critical"
       >
         <Icon source={PaintBrushMajor} tone="base" />
       </Box>
@@ -157,7 +135,7 @@ export function WithOutline() {
 
 export function WithBorderRadius() {
   return (
-    <Box background="bg" padding="400" borderRadius="2">
+    <Box background="bg-surface" padding="400" borderRadius="200">
       <Icon source={PaintBrushMajor} tone="info" />
     </Box>
   );
@@ -167,44 +145,44 @@ export function WithResponsivePadding() {
   return (
     <BlockStack gap="400">
       <Box
-        background="bg"
+        background="bg-surface"
         padding={{xs: '200', sm: '800'}}
-        borderWidth="1"
+        borderWidth="025"
         borderColor="border"
       >
         <Icon source={PaintBrushMajor} tone="base" />
       </Box>
       <Box
-        background="bg"
+        background="bg-surface"
         padding="200"
         paddingBlockStart={{xs: '400', sm: '1000'}}
-        borderWidth="1"
+        borderWidth="025"
         borderColor="border"
       >
         <Icon source={PaintBrushMajor} tone="base" />
       </Box>
       <Box
-        background="bg"
+        background="bg-surface"
         padding="200"
         paddingBlockEnd={{xs: '400', sm: '1000'}}
-        borderWidth="1"
+        borderWidth="025"
         borderColor="border"
       >
         <Icon source={PaintBrushMajor} tone="base" />
       </Box>
       <Box
-        background="bg"
+        background="bg-surface"
         padding="200"
         paddingInlineStart={{xs: '400', sm: '1000'}}
-        borderWidth="1"
+        borderWidth="025"
         borderColor="border"
       >
         <Icon source={PaintBrushMajor} tone="base" />
       </Box>
       <Box
-        background="bg"
+        background="bg-surface"
         paddingInlineEnd={{xs: '400', sm: '1000'}}
-        borderWidth="1"
+        borderWidth="025"
         borderColor="border"
       >
         <Icon source={PaintBrushMajor} tone="base" />