+++ polaris-react/src/utilities/index-filters/types.ts
@@ -1,5 +1,0 @@
-export enum IndexFiltersMode {
-  Default = 'DEFAULT',
-  Filtering = 'FILTERING',
-  EditingColumns = 'EDITING_COLUMNS',
-}