+++ polaris.shopify.com/src/components/TokenList/TokenList.tsx
@@ -10,6 +10,7 @@ import Link from 'next/link';
 import {useGlobalSearchResult} from '../GlobalSearch/GlobalSearch';
 import {ClipboardMinor} from '@shopify/polaris-icons';
 import SearchResultHighlight from '../SearchResultHighlight';
+import {toPxs} from '@shopify/polaris-tokens';
 
 interface ColumnsConfig {
   preview: boolean;
@@ -23,7 +24,7 @@ const defaultColumnsConfig: ColumnsConfig = {
   preview: true,
   name: true,
   value: true,
-  figmaUsage: true,
+  figmaUsage: false,
   description: true,
 };
 
@@ -50,9 +51,9 @@ function TokenList({
             <thead>
               <tr>
                 {columns.preview && <th></th>}
-                {columns.name && <th>Token name</th>}
-                {columns.value && <th>Current value</th>}
-                {columns.figmaUsage && <th>Figma usage</th>}
+                {columns.name && <th>Name</th>}
+                {columns.value && <th>Value</th>}
+                {columns.figmaUsage && <th>Figma</th>}
                 {columns.description && <th>Description</th>}
               </tr>
             </thead>
@@ -90,7 +91,7 @@ function getFigmaUsageForToken(
   const REM = 16;
 
   if (value.startsWith('rgba')) {
-    usage = figmaColorNames[name] ? `Use ${figmaColorNames[name]}` : '-';
+    usage = figmaColorNames[name] ? `Use ${figmaColorNames[name]}` : '—';
   } else if (name.startsWith('shadow')) {
     usage = 'Use shadow styles from UI kit';
   } else if (name.includes('breakpoint')) {
@@ -206,7 +207,7 @@ function TokenListItem({
               </span>
             </td>
           )}
-          {columns.value && <td className={styles.Value}>{value}</td>}
+          {columns.value && <td className={styles.Value}>{toPxs(value)}</td>}
           {columns.figmaUsage && (
             <td className={styles.FigmaUsage}>{figmaUsage || '—'}</td>
           )}
@@ -308,8 +309,8 @@ function TokenPreview({name, value}: TokenPreviewProps) {
     );
   }
 
-  // Spacing
-  else if (name.includes('space')) {
+  // Spacing and Width
+  else if (name.includes('space') || name.includes('width')) {
     return (
       <div
         {...previewDivAttributes}
@@ -342,6 +343,53 @@ function TokenPreview({name, value}: TokenPreviewProps) {
             background: 'var(--text)',
           }}
         ></div>
+      </div>
+    );
+  }
+
+  // Height
+  else if (name.includes('height') && !name.includes('line-height')) {
+    return (
+      <div
+        {...previewDivAttributes}
+        style={{
+          display: 'flex',
+          alignItems: 'flex-start',
+        }}
+      >
+        <div
+          {...previewDivAttributes}
+          style={{
+            display: 'flex',
+            flexDirection: 'column',
+            alignItems: 'center',
+          }}
+        >
+          <div
+            style={{
+              aspectRatio: '1/1',
+              borderRadius: 100,
+              minHeight: '10px',
+              background: 'var(--text)',
+            }}
+          ></div>
+          <div
+            style={{
+              minWidth: '30px',
+              height: value,
+              background: 'var(--text)',
+              opacity: 0.15,
+            }}
+          ></div>
+          <div
+            style={{
+              aspectRatio: '1/1',
+              borderRadius: 100,
+              minHeight: '10px',
+              background: 'var(--text)',
+            }}
+          ></div>
+        </div>
       </div>
     );
   }
@@ -408,6 +456,22 @@ function TokenPreview({name, value}: TokenPreviewProps) {
         Hello
         <br />
         World
+      </div>
+    );
+  }
+
+  // Letter spacing
+  else if (name.includes('letter-spacing')) {
+    return (
+      <div
+        {...previewDivAttributes}
+        style={{
+          display: 'flex',
+          letterSpacing: value,
+          background: 'transparent',
+        }}
+      >
+        Entrepreneurs
       </div>
     );
   }